stream: light-weight buffer recycling while in new_buffer
When recycle_buffer() is called while in the new_buffer event handler,
don't send the reuse_buffer event. Instead, mark the buffer for
recycling by putting the buffer ID in the IO area, which is more
light-weight. When need_input reaches the server, it will recycle the
buffer.

Also introduce a helper function for sending the reuse_buffer event.

Change-Id: I900e75694efce2fa7e12840eaf53a7f6b7ae7e8a