build: Fix dependency objects for system libraries
The dependency checks for system libraries is not really portable.

Not all C standard libraries have separate rt and math libraries, so the
requirement should be set to false.

Additionally, finding threading libraries should be left to Meson
itself, using the `dependency('threads')` object, which will do the
right thing depending on platform and compiler.