Install header needed to compile pipewire-jack
I'm trying to compile pipewire-jack as a standalone module instead of getting it into the full pipewire source and recompile everything. This header is needed to be installed in order for the include preprocessor checks to pass. This is the only change needed on this side apparently, everything else needs to be done in the pipewire-jack side and it seems to be quite a bit more stuffy.