alsa: clamp the period size to something sensible
some clients (sweep) want to negotiate the smallest possible period
size we can support, which is expressed in bytes and independent
of samplerate, format and number of channels.

For files with high samplerate and channelcount, this results in periods
of 8 frames, which is too small to handle.

Instead round up the period size to something we can likely handle,
taking into account the samplerate and number of channels.