link: only reuse buffers if the port is at least in paused state
If existing buffers of a port are reused during allocation then this port
is not changed in any way. If it is still in the 'ready' state then will
remains there and the link activation never finishes.
This can happen if a previous link setup was aborted during allocation.

Fix this by only reusing buffer if the port is at least paused.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>