gst/pwaudioringbuffer: request pause/play on the appropriate stream state changes
This allows the client to properly go to PAUSED when the session manager
unlinks the stream and go again to PLAYING when the sm re-links it.
This allows the session manager to implement policies without letting
the client pipeline freeze (in the absence of a running audio clock)
when it is unlinked. Note that in case the client doesn't handle the
request, there is still no issue. Like in pulseaudio, the clock just
freezes, so the pipeline stops progressing.

This is similar to the pulseaudio cork/uncork mechanism.